refactor application with helper script
